Implemented responding to requests

+++ core/input.php
@@ -115,13 +115,81 @@
    // CANCEL DATE REQUEST
    } else if (post::isset('date_cancel')) {
       
-      if (post::val('confirm') == 'Confirm') {
+      if (post::val('confirm') == 'Confirm' && !empty(post::val('from'))) {
          
          db::instance()->action(
             'DELETE FROM requests WHERE fromUser = ? AND toUser = ?',
             array(
                session::username(),
                strtolower(get::val('username'))
+            )
+         );
+
+         misc::redirect('self');
+      }
+
+   // ACCEPT REQUEST
+   } else if (post::isset('accept')) {
+         
+      if (post::val('accept_confirm') == 'Accept' && !empty(post::val('from'))) {
+         
+         db::instance()->action(
+            'DELETE FROM requests WHERE fromUser = ? AND toUser = ?',
+            array(
+               post::val('from'),
+               session::username()
+            )
+         );
+
+         db::instance()->action(
+            'INSERT INTO dates (code, person1, person2, date) VALUES (?, ?, ?, ?)',
+            array(
+               misc::generateDate(),
+               session::username(),
+               post::val('from'),
+               date('Y-m-d H:i:s')
+            )
+         );
+
+         misc::redirect('self');
+      }
+
+   // ACCEPT REQUEST
+   } else if (post::isset('accept')) {
+         
+      if (post::val('accept_confirm') == 'Accept' && !empty(post::val('from'))) {
+         
+         db::instance()->action(
+            'DELETE FROM requests WHERE fromUser = ? AND toUser = ?',
+            array(
+               post::val('from'),
+               session::username()
+            )
+         );
+
+         db::instance()->action(
+            'INSERT INTO dates (code, person1, person2, date) VALUES (?, ?, ?, ?)',
+            array(
+               misc::generateDate(),
+               session::username(),
+               post::val('from'),
+               date('Y-m-d H:i:s')
+            )
+         );
+
+         misc::redirect('self');
+      }
+
+   // DECLINE REQUEST
+   } else if (post::isset('decline')) {
+            
+      if (post::val('decline_confirm') == 'Decline' && !empty(post::val('from'))) {
+         
+         db::instance()->action(
+            'DELETE FROM requests WHERE fromUser = ? AND toUser = ?',
+            array(
+               post::val('from'),
+               session::username()
             )
          );
